Auto stash before merge of "expandTesting" and "origin/master"
Former-commit-id: 2cbb5b234dbf438214972b911bbe6263a454d27e

+++ id/src/androidTest/java/com/simprints/id/Utils.java
@@ -2,6 +2,7 @@ package com.simprints.id;
 
 import android.content.Context;
 
+import com.simprints.cerberuslibrary.RealmUtility;
 import com.simprints.remoteadminclient.Configuration;
 import com.simprints.remoteadminclient.api.DefaultApi;
 import com.simprints.remoteadminclient.auth.ApiKeyAuth;
@@ -9,6 +10,8 @@ import com.squareup.okhttp.OkHttpClient;
 
 import java.io.File;
 import java.util.concurrent.TimeUnit;
+
+import io.realm.RealmConfiguration;
 
 class Utils {
 
@@ -30,11 +33,14 @@ class Utils {
         return apiInstance;
     }
 
-    static void clearApplicationData(Context context) {
+    static void clearApplicationData(Context context, RealmConfiguration realmConfiguration) {
+        clearRealmDatabase(realmConfiguration);
         File cacheDirectory = context.getCacheDir();
+        if (cacheDirectory == null) return;
         File applicationDirectory = new File(cacheDirectory.getParent());
         if (applicationDirectory.exists()) {
             String[] fileNames = applicationDirectory.list();
+            if (fileNames == null) return;
             for (String fileName : fileNames) {
                 if (!fileName.equals("lib")) {
                     deleteFile(new File(applicationDirectory, fileName));
@@ -58,4 +64,8 @@ class Utils {
         return deletedAll;
     }
 
+    private static void clearRealmDatabase(RealmConfiguration realmConfiguration) {
+        if (realmConfiguration == null) return;
+        new RealmUtility().clearRealmDatabase(realmConfiguration);
+    }
 }
